More ADS and dentry alignment and padding issues
So... I think that ALL the filename fields (short name, file name, stream name)
actually have a 2-byte null terminator, but NOT if that field has zero length.
This is now taken into account.

A warning is now issued if the dentry length is unexpectedly large.

We will now write all dentries with the "canonical" length regardless of what
dentry length was read in (as we are throwing out whatever junk was appended to
it anyway...).