Build fixes
Use casts to or from `uintptr_t' to avoid warnings on 32-bit platforms when we
get or set the FUSE private data in mount.c.

Make the XATTR interface optional.  Support is omitted automatically if the
setxattr() function or the `attr/xattr.h' header is not found.

Fix a couple of the test cases.