to_leXX() -> cpu_to_leXX(), leXX_to_cpu()
Split the macros to_leXX() into separate macros for converting *from* in-memory
little endian *to* the correct in-CPU value, and for converting *from* a correct
in-CPU value *to* in-memory little-endian.  In reality these macros are defined
the same, but it makes sense to distinguish the cases (as is done in other
projects such as the Linux kernel).

By the way, wimlib still has NOT actually been tested on big-endian systems.