Metadata resource compatibility
For compatibility with 7zip, do not write any extra random bytes to the end of
the metadata resources.

This means that without adding some other source of random output, adding
identical images will result in identical metadata resources.  This is still
allowed, but for compatibility with ImageX, lookup table entries are duplicated
rather than shared.