mount_image.c:  Implement versioned unmount messages
The ad-hoc format of the unmount messages has been expanded with a header that
specifies the message size and the minimum and current version of the library
sending the message.  This could potentially make things go smoother if the
messages are changed and unmounts are done with a different library version than
the mount is with.

Also, some code was refactored to reduce function length.