NTFS capture:  Get DOS names from ntfs_readdir()
Instead of calling ntfs_get_ntfs_dos_name(), the behavior of which was changed
in NTFS-3g v2013.1.13 and no longer works for files with multiple hard links,
one of which is associated with a hard name, just remember the DOS names when
they're returned from the ntfs_readdir() callback.  This works by building a
temporary map from NTFS inode numbers to DOS names each time a directory is
captured.