WIM capture:  Share inodes immediately
Maintain a map (inode number, device number) => (WIM inode) throughout the
process of capturing a WIM image so that we can immediately detect a dentry that
refers to an already-captured inode.  Then, the inode can be shared immediately,
and its streams need not be read again.  In addition, the inodes need not be
sent through more complicated the hard link group disambiguation code; also, on
UNIX, capturing files with the same inode number but on different devices now
guaranteed to work correctly.