New compression/decompression API
To avoid the proliferation of functions for compressing and decompressing
in different formats, allow all the compression algorithms to be accessed
using a single API:

Compression:
	- wimlib_create_compressor()
	- wimlib_compress()
	- wimlib_free_compressor()
	- wimlib_set_default_compressor_params()

Decompression:
	- wimlib_create_decompressor()
	- wimlib_decompress()
	- wimlib_free_decompressor()
	- wimlib_set_default_decompressor_params()

This also makes it easier to allocate larger blocks of memory or do other
initializations in any decompressor or compressor implementation.

This commit adds a skeleton for the LZMS compressor but it doesn't do
anything yet.