LZMS:  Do not do x86 translations starting in last 16 bytes
The value used before (last 11 bytes) was wrong; based on tests with the
MS implementation no translation can occur following an opcode starting
in the last 16 bytes of a data block.

Tested with two different length opcodes to make sure that it's the start
of the opcode that matters, not the start of the translated address.