init_upcase(): Fix and cleanup
The previous code used the wrong array size and consequently left 0's in
entries 0x8000 and above.  This would incorrectly cause all characters in
that range to be considered equal to each other case-insensitively.  Fix
it and clean up the code.

Note: the code still does not do real UTF-16 case folding.