win32_capture.c: Ignore unnamed data stream of reparse point
This data will not be archived; only the reparse data will.  So there's
no need to create a 'struct wim_lookup_table_entry' for it.

If a reparse point had an unnamed data stream, the assertion introduced
in commit 321750354891a0968ca0c3664417eae73b9414aa would be triggered.
This is just as well, because this case wasn't being handled in the most
logical way.