Align stack when entering wimlib on 32-bit x86
This may resolve https://wimlib.net/forums/viewtopic.php?t=669.  The
reported crash happened in sha1_blocks_x86_avx_bmi2() on the first
instruction that writes to the stack with vmovqda.  That suggests the
stack was misaligned.  If this indeed fixes the bug, then it has
actually always existed but it got exposed by new code in v1.14.0.