Add a clang-format file
This approximates the de facto coding style of this project, which is
quite similar to the Linux kernel coding style.

This is just in case people to use it with 'git clang-format', as
requested at https://wimlib.net/forums/viewtopic.php?t=666.  I'm not
planning to reformat all the existing code to make it match exactly.